% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BHRecruitFun.r
\name{BHRecruitFun}
\alias{BHRecruitFun}
\title{Calculates the recruitment from the Beverton-Holt SRR model.}
\usage{
BHRecruitFun(currEggProd, steepness, R0, UnfishedEggProd, RecDev = 1)
}
\arguments{
\item{currEggProd}{Current egg production - total number of eggs or spawning biomass at the
beginning of the year.}

\item{steepness}{Steepness (h) of the stock-recruit relationship.}

\item{R0}{Number of recruits at unfished equilibrium.}

\item{UnfishedEggProd}{Egg production (or spawning biomass) in the equilibrium unfished state.}

\item{RecDev}{Log-normally distributed recruitment deviation.}
}
\value{
To add details later.
}
\description{
Beverton-Holt Stock Recruit Relationship.
}
\author{
Adrian Hordyk
}
\seealso{
\code{\link{}}
}

