% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DynamicMod_AgeStuc.r
\name{DynamicMod_AgeStuc}
\alias{DynamicMod_AgeStuc}
\title{Projection Model}
\usage{
DynamicMod_AgeStuc(SimPars, RecDevVec)
}
\arguments{
\item{SimPars}{An object of class \code{list} that contains all parameters
required to run GTG model.  Full description of model to be added at later
date.}

\item{RecDevVec}{A vector of log-normally distributed recruitment deviations of length equal to
number of projection years (to be added to SimPars later).}
}
\value{
To add details later.
}
\description{
Age-structured GTG dynamic projection model.
of stock and relative yield. This model is parameterised with the life-history ratios.
}
\author{
Adrian Hordyk
}
\seealso{
\code{\link{}}
}

