% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GenSPRYPR.R
\name{GenSPRYPR}
\alias{GenSPRYPR}
\title{Generate size structure using GTG model}
\usage{
GenSPRYPR(SimPars)
}
\arguments{
\item{SimPars}{An object of class \code{list} that contains all parameters
required to run GTG model.  Full description of model to be added at later
date.}
}
\value{
To add details later.
}
\description{
Function to simulate length-structured growth-type-group (GTG) model to
generate size equilibrium composition of population and catch, as well as SPR
of stock and relative yield.  More details to be added.
}
\examples{
\dontrun{

}
}
\author{
Adrian Hordyk
}
\seealso{
\code{\link{}}
}

