% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LoadAssessPars.R
\name{LoadAssessPars}
\alias{LoadAssessPars}
\title{Load LBSPR Assessment Parameters}
\usage{
LoadAssessPars(PathtoAssessFile = "~/PathToAssessFile",
  AssessParFileName = "AssessPars", AssessParExt = ".csv", ind = 1,
  LenMids, IncludeDatFile = FALSE, LenDatType = list("raw", "comp"),
  LHR_OverRide = list())
}
\arguments{
\item{PathtoAssessFile}{an object of class \code{character} containing the
full path to the Assessment Parameter file}

\item{AssessParFileName}{an object of class \code{character} containing the
name of Assessment Parameter file}

\item{AssessParExt}{an object of class \code{character} containing the file
extension of Assessment Parameter file (default is ".csv" and currently can
not handle anything else)}

\item{ind}{an optional \code{numeric} value indicating which column contains
the parameters (default is 1).}

\item{LenMids}{A vector of the midpoints of the length classes for the observed length data.}
}
\value{
Returns a list of assessment parameters (\code{AssessPars}) that is
  used in other functions.
}
\description{
This function reads in a csv file that contains the parameters required for
running the LB-SPR Assessment model.  See LoadAssessPars ReadMe for details
of contents of SimParFile.
}
\author{
Adrian Hordyk
}

