% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LoadSimPars.r
\name{LoadSimPars}
\alias{LoadSimPars}
\alias{LoadSimParsTEST}
\title{Load LBSPR Simulation Parameters}
\usage{
LoadSimParsTEST(PathtoSimFile = "~/PathToSimFile",
  SimParFileName = "SimData", SimParExt = ".csv", ind = 1)
}
\arguments{
\item{PathtoSimFile}{an object of class \code{character} containing the full path to the Simulation Parameter file}

\item{SimParFileName}{an object of class \code{character} containing the name of Simulation Parameter file}

\item{SimParExt}{an object of class \code{character} containing the file extension of Simulation Parameter file (default is ".csv")}

\item{ind}{an optional \code{numeric} value indicating which column contains the parameters (default is 1)}
}
\value{
Returns a list of simulation parameters
}
\description{
This function reads in a csv file that contains the parameters for simulating the LB-SPR model.  See SimData ReadMe for details of
contents of SimParFile.
}
\author{
Adrian Hordyk
}

