% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MakeLengthComp.R
\name{MakeLengthComp}
\alias{MakeLengthComp}
\alias{MakeLengthCompFun}
\title{Create Length Frequency Composition}
\usage{
MakeLengthComp(PathtoLenDat, LenDatFileName, LenDatExt = ".csv",
  Header = FALSE, Linc = 5, AssessPars, Multi = 1.25, sep = "period")
}
\arguments{
\item{PathtoLenDat}{An object of class \code{character} that contains the
full path to the raw length data file.}

\item{LenDatFileName}{An object of class \code{character} that contains the
name of the length data file.}

\item{LenDatExt}{An object of class \code{character} that contains the
extension of the length data file (default is ".csv" and currently does not
handle anything else).}

\item{Header}{An \code{logical} indicating if the first row of the CSV file
is a header (Default is FALSE).}

\item{Linc}{An \code{numeric} value indicating the width of the length class
(in same units as \code{AssessPars}).}

\item{AssessPars}{A \code{list} containing the parameters required for the
LB-SPR assessment. Created by \code{LoadAssessPars} function.}

\item{Multi}{An \code{numeric} object that specifies the maximum length class as a mulitiple of \code{AssessPars$Linf}.}

\item{sep}{Decimals are indicated by \code{"period"} (default) or \code{"comma"}}
}
\value{
A \code{list} containing the Length Bins (\code{LenBins}), the Length
  Frequency (\code{LenFreq}) and the midpoints of the length bins
  (\code{LenMids}).
}
\description{
\code{MakeLengthCompFun} is used to create a length frequency composition
from raw length data.
}
\author{
Adrian Hordyk
}

