% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PredictMSlope.r
\name{PredictMSlope}
\alias{PredictMSlope}
\title{Predict the value of Mslope parameter from the LM and Pars.}
\usage{
PredictMSlope(Pars)
}
\arguments{
\item{Pars}{An object of class \code{list} that contains all parameters
required to run GTG model.}
}
\description{
This function Predict the value of Mslope parameter from the LM and Pars.
}
\author{
Adrian Hordyk
}

