% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimMod_AgeEq.r
\name{SimMod_AgeEq}
\alias{SimMod_AgeEq}
\title{Generate size structure using GTG model and calculate relative YPR and SPR}
\usage{
SimMod_AgeEq(SimPars)
}
\arguments{
\item{SimPars}{An object of class \code{list} that contains all parameters
required to run GTG model.  Full description of model to be added at later
date.}
}
\value{
To add details later.
}
\description{
Function to simulate age-structured growth-type-group (GTG) model to
generate size equilibrium composition of population and catch, as well as SPR
of stock and relative yield.
}
\author{
Adrian Hordyk
}

