% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimMod_LenStrucRate.r
\name{SimMod_LenStrucRate}
\alias{SimMod_LenStrucRate}
\title{Generate size structure using GTG model and calculate relative YPR and SPR}
\usage{
SimMod_LenStrucRate(SimPars, Mpar, ...)
}
\arguments{
\item{SimPars}{An object of class \code{list} that contains all parameters
required to run GTG model.  Full description of model to be added at later
date.}
}
\value{
To add details later.
}
\description{
Function to simulate length-structured growth-type-group (GTG) model to
generate size equilibrium composition of population and catch, as well as SPR
of stock and relative yield. This model is follows the same length-structured GTG
approach of \code{SimMod_LHR} but uses the rate parameters directly (i.e M, K, F)
rather than the ratios. The results from the two models should be identical if the
life history and fishing ratios are the same. This model is mainly for testing purposes
and not likely to be used for simulation.  Use \code{SimMod_AgeEq} for an age-structured simulation model.
}
\author{
Adrian Hordyk
}

