% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimParCalcs.r
\name{SimParsCalc}
\alias{SimParsCalc}
\title{Calculate values for derived simulations parameters.}
\usage{
SimParsCalc(SimPars, ModType = "Len")
}
\arguments{
\item{SimPars}{An object of class \code{list} that contains all parameters
required to run GTG model. This function calculates additional parameters and adds them to \code{SimPars}.}

\item{ModType}{A character string of either \code{Len} or \code{Age} to indicate if the equilibrium simulation model is length or age structured respectively.}
}
\value{
Returns a list of simulation parameters
}
\description{
This function to be run after \code{LoadSimPars} and calculates the derived parameters. Re-run this
function each time of the \code{SimPars} parameters is manually changed.
}
\author{
Adrian Hordyk
}

