% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SingleGTGDynamic.r
\name{SingleGTGDynamic}
\alias{SingleGTGDynamic}
\title{Single GTG Dynamic Projection Model Function}
\usage{
SingleGTGDynamic(AgeVec, LinfGTG, MparGTG, RecGTGMonth, TSkpar, L50gtg, L95gtg,
  SL50, SL95, Walpha, Wbeta, FecB, LenMids, LenBins, Mpow, FparYr, MeanLinf,
  LastNAgeVec, LastNAgeUFVec, TrackRecruitsGTG, TS, RecGTGMonthVec, RecProb, R0,
  RecGTGMonthProb)
}
\arguments{
\item{This}{function is called by other functions and parameters do not need to be set.}
}
\description{
Function for running age-structured GTG model. Not meant to be run.
}
\author{
Adrian Hordyk
}
\seealso{
\code{\link{}}
}

