% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SingleYearEqSim.r
\name{SingleYearEq}
\alias{SingleYearEq}
\title{Single Year Equilibrium Projection Model}
\usage{
SingleYearEq(FparYr, SimPars)
}
\arguments{
\item{FparYr}{Fishing mortality (in monthly units of time). Usually set to 0 for unfished equlibrium.}

\item{SimPars}{An object of class \code{list} that contains all parameters
required to run GTG model.  Full description of model to be added at later
date.}
}
\value{
To add details later.
}
\description{
Equilibrium model to project a single year of the age-structured GTG model (usually first unfished year)
}
\author{
Adrian Hordyk
}
\seealso{
\code{\link{}}
}

