% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/downloadADMB.r
\name{CopyADMBFile}
\alias{CopyADMBFile}
\title{Download LBSPR executable.}
\usage{
CopyADMBFile(System = "Win", DestPath = NULL)
}
\arguments{
\item{System}{A \code{character} string of either "Win" for Windows OS compiled executable, or anything else (not blank) for UNIX OS}

\item{DestPath}{A code{character} string detailing the full path where the binary file is to be downloaded.}
}
\description{
Function to download the LBSPR binary from online repository and copy to local machine.
}
\author{
Adrian Hordyk
}

