% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PredictKSlope.r
\name{PredictKSlope}
\alias{PredictKSlope}
\title{Predict the value of kslope parameter from the LM and Pars.}
\usage{
PredictKSlope(Pars)
}
\arguments{
\item{Pars}{An object of class \code{list} that contains all parameters
required to run GTG model.}
}
\description{
This function Predict the value of kslope parameter from the LM and Pars.
}
\author{
Adrian Hordyk
}

